+++ reservationservice/src/main/java/com/epam/reservation/controller/ReservationController.java
@@ -1,4 +1,4 @@
-package com.epam.reservationservice.controller;
+package com.epam.reservation.controller;
 
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -7,7 +7,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 @RequestMapping("/v1/api")
-public interface ReservationServiceControllerInterface {
+public interface ReservationController {
 
 	@PostMapping("/reservations")
 	public String addReservation();
