+++ reservationservice/src/main/java/com/epam/reservation/service/ReservationService.java
@@ -1,5 +1,5 @@
-package com.epam.reservationservice.service;
+package com.epam.reservation.service;
 
-public class ReservationServiceClass {
+public class ReservationService {
 
 }
