+++ reservationservice/src/test/java/com/epam/reservation/ReservationApplicationTest.java
@@ -1,16 +1,16 @@
-package com.epam.reservationservice;
+package com.epam.reservation;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
 import org.springframework.boot.test.web.client.TestRestTemplate;
 
-import com.epam.reservationservice.entity.ReservationDetails;
+import com.epam.reservation.entity.ReservationDetails;
 
 
 
 @SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
-class ReservationserviceApplicationTests {
+class ReservationApplicationTest {
 
 	TestRestTemplate restTemplate = new TestRestTemplate();
